Make intermediate commit that uses defaultdict of strings to show which sequences of lines occur in input
get_abc.py now creates a string of three characters in each call to make_output_line(). Each character
represents a type of line: '' is represented by 'A', 'Week of ...' by 'B', '=====...' by 'C', etc.
The three-character strings are inserted into a defaultdict, which is output at the end of the script.

On branch inside_work_on_transform
Changes to be committed:
    modified:   get_abc.py