Make another intermediate commit that tracks the *type* of action within the three-character strings
Instead of using 'E' for an action line in the input, this version of get_abc.py uses 'b', 's', or 'w'.
The action line itself is also written as 'act=n', where n is in 'bsw'. ('b': bedtime, 's': sleep,
'w': wake).

On branch inside_work_on_transform
Changes to be committed:
    modified:   get_abc.py