Make do_transform output full NIGHTs and NAPs, though it misses >1 edge case
1) When the last NAP(s) of a NIGHT occur at or after midnight of the following day,
the value displayed for the start time of the NAP(s) does *not* indicate that the
start time is in relation to the newer day.

2) In do_transform.py, function get_duration() adjusts the hours and the minutes
values completely separately, which can produce a wrong value for the duration.
See e.g., the first NAP after 'NIGHT, 2017-03-02, 23:15', which shows a value of
2.75 for the duration when the correct value is 1.75.

On branch work_on_transform
Changes to be committed:
    modified:   src/transform/do_transform.py
    deleted:    xtraneous/output_mk_processes_04.txt