Create workaround for edge case 1) in commit 24cf00
Alter db so that the sl_nap table has a SERIAL p k. This ensures that, if naps are entered into
the db in chronological order, an earlier nap will have a lower p k value, regardless of the
start value of the nap.

On branch work_on_transform
Changes to be committed:
    modified:   db/create_tables.sql