Keep working on the non-class version of do_transform.py
Am having success with this; it may wind up being the version I choose.
It's a question of picking either a class that will end up being
a singleton, or a non-class implementation that uses a closure
but is a bit overlong (readability may suffer).

On branch work_on_old_do_transform
Changes to be committed:
    new file:   xtraneous/do_transform_no_class/do_transform.py
    new file:   xtraneous/do_transform_no_class/tests/test_do_transform.py